Create Class Student
Create a class called "Student" with the following private member variables:
•	name (string)
•	age (int)
•	grade (int)
The class should have the following public member functions:
•	A constructor that sets the name, age, and grade.
•	A function called "getName" that returns the name of the student.
•	A function called "getAge" that returns the age of the student.
•	A function called "getGrade" that returns the grade of the student.
•	A function called "setGrade" that sets the grade of the student.
In the main function, create an instance of the Student class and set its name, age, and grade. Then use the setGrade function to change the grade of the student. Print the name, age, and grade of the student using the getName, getAge, and getGrade functions.